+++ src/main/java/com/tm/calemieconomy/util/IItemCurrencyHolder.java
@@ -4,11 +4,11 @@ import net.minecraft.world.item.ItemStack;
 
 public interface IItemCurrencyHolder {
 
-    int getCurrency(ItemStack stack);
-    int getCurrencyCapacity();
-    void setCurrency(ItemStack stack, int amount);
-    boolean canDepositCurrency(ItemStack stack, int amount);
-    boolean canWithdrawCurrency(ItemStack stack, int amount);
-    void depositCurrency(ItemStack stack, int amount);
-    void withdrawCurrency(ItemStack stack, int amount);
+    long getCurrency(ItemStack stack);
+    long getCurrencyCapacity();
+    void setCurrency(ItemStack stack, long amount);
+    boolean canDepositCurrency(ItemStack stack, long amount);
+    boolean canWithdrawCurrency(ItemStack stack, long amount);
+    void depositCurrency(ItemStack stack, long amount);
+    void withdrawCurrency(ItemStack stack, long amount);
 }
